+++ changelog.md
@@ -1,3 +1,3 @@
 ### General
 
-- ***Added*** carpet rule `commandAddressTp`
+- ***Added*** carpet rule `fakePlayerInventoryFix` to fix [carpet-2110](https://github.com/gnembon/fabric-carpet/issues/2110) for 1.21.9-10